news-summary-backend/requirements.txt
flask
flask-cors
pytesseract
pdfplumber
Pillow
transformers
torch
